Se corrigio funcionalidad de botones editar,borrar,crear de la pagina de comidas.Se agregaron restricciones para que solo si sos superuser puedas ver y editar,borrar,crear las comidas

+++ Comidas/views.py
@@ -6,51 +6,8 @@ from Comidas.forms import Formulario_comidas
 from django.contrib.auth.decorators import login_required
 
 
-@login_required
-def crear_comida(request):
-    if request.user.is_superuser:
-        if request.method == 'POST':
-            form = Formulario_comidas(request.POST, request.FILES)
 
-            if form.is_valid():
-                Comidas.objects.create(
-                    name = form.cleaned_data['name'],
-                    description = form.cleaned_data['description'],
-                    price = form.cleaned_data['price'],
-                    image = form.cleaned_data['image']
-                )
-                
-                return redirect(lista_comidas)
 
-        elif request.method == 'GET':
-            form = Formulario_comidas()
-            context = {'form':form}
-            return render(request, 'Comidas/nueva_comida.html', context=context)
-    return redirect('login')
-
-def actualizar_comidas(request, pk):
-    if request.method == 'POST':
-        form = Formulario_comidas(request.POST)
-        if form.is_valid():
-            comida = Comidas.objects.get(id=pk)
-            comida.name = form.cleaned_data['name']
-            comida.price = form.cleaned_data['price']
-            comida.description = form.cleaned_data['description']
-            comida.stock = form.cleaned_data['stock']
-            comida.save()
-
-            return redirect(lista_comidas)
-
-    elif request.method == 'GET':
-        comida = Comidas.objects.get(id=pk)
-
-        form = Formulario_comidas(initial={
-                                        'name':comida.name,
-                                        'price':comida.price, 
-                                        'description':comida.description,
-                                        'stock':comida.stock})
-        context = {'form':form}
-        return render(request, 'Comidas/actualizar_comida.html', context=context)
 
 
 
@@ -64,7 +21,6 @@ def lista_comidas(request):
             return render(request, 'Comidas/lista_comidas.html', context=context)
 
     
-
 def primer_formulario(request):
     print(request.method)
     if request.method == 'POST':
@@ -77,12 +33,81 @@ def buscar_comida(request):
     context = {'comida':comida}
     return render(request, 'Comidas/buscar_comida.html', context=context)
 
+
+
+
+def detalle_comidas(request,pk):
+    if request.user.is_authenticated:
+        if request.method == 'GET':
+            comida = Comidas.objects.get(pk=pk)
+            context = {'comida':comida}
+            return render(request, 'Comidas/detalle_comidas.html', context=context)
+    return redirect('login')
+
+
+def crear_comida(request):
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            if request.method == 'POST':
+                form = Formulario_comidas(request.POST, request.FILES)
+
+                if form.is_valid():
+                    Comidas.objects.create(
+                        name = form.cleaned_data['name'],
+                        description = form.cleaned_data['description'],
+                        price = form.cleaned_data['price'],
+                        
+                        image = form.cleaned_data['image']
+                    )
+                    
+                    return redirect(lista_comidas)
+
+            elif request.method == 'GET':
+                form = Formulario_comidas()
+                context = {'form':form}
+                return render(request, 'Comidas/nueva_comida.html', context=context)
+        return redirect('login')
+    return redirect('login')
+
+def actualizar_comidas(request, pk):
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            if request.method == 'POST':
+                form = Formulario_comidas(request.POST)
+                if form.is_valid():
+                    comida = Comidas.objects.get(id=pk)
+                    comida.name = form.cleaned_data['name']
+                    comida.price = form.cleaned_data['price']
+                    comida.description = form.cleaned_data['description']
+                    
+                    comida.save()
+
+                    return redirect(lista_comidas)
+
+            elif request.method == 'GET':
+                comida = Comidas.objects.get(id=pk)
+
+                form = Formulario_comidas(initial={
+                                                'name':comida.name,
+                                                'price':comida.price, 
+                                                'description':comida.description,
+                                                })
+                context = {'form':form}
+                return render(request, 'Comidas/actualizar_comidas.html', context=context)
+        return redirect('login')
+    return redirect('login')
+
+
 def borrar_comida(request, pk):
-    if request.method == 'GET':
-        comida = Comidas.objects.get(pk=pk)
-        context = {'comida':comida}
-        return render(request, 'Comidas/borrar_comidas.html', context=context)
-    elif request.method == 'POST':
-        comida = Comidas.objects.get(pk=pk)
-        comida.delete()
-        return redirect(lista_comidas)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+            if request.method == 'GET':
+                comida = Comidas.objects.get(pk=pk)
+                context = {'comida':comida}
+                return render(request, 'Comidas/borrar_comidas.html', context=context)
+            elif request.method == 'POST':
+                comida = Comidas.objects.get(pk=pk)
+                comida.delete()
+                return redirect(lista_comidas)
+        return redirect('login')
+    return redirect('login')
